LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/referrer-policy-from-rules.https_6-6-expected.txt


FAIL treat legacy referrer policy values as invalid assert_equals: must not be prefetched expected (undefined) undefined but got (string) "prefetch"
